requirements.txt
streamlit
pandas
plotly
google-generativeai
Pillow
kaleido
python-dotenv
